limited genplus_learning models and serializer to the specific app
