+++ openedx/features/genplus_features/genplus/api/v1/urls.py
@@ -23,6 +23,5 @@ urlpatterns = [
     url(r'^userinfo/$', UserInfo.as_view()),
     url(r'^characters/select/(?P<pk>\d+)/$', CharacterViewSet.as_view({"put": "select_character"})),
     url(r'^classes/add_class/(?P<group_id>\w+)/$', ClassViewSet.as_view({"put": "add_my_class"})),
-    url(r'^classes/lesson/unlock/(?P<pk>\d+)/$', ClassViewSet.as_view({"put": "unlock_lesson"})),
     path('', include(router.urls)),
 ]
