+++ openedx/features/genplus_features/genplus/api/v1/views.py
@@ -1,4 +1,3 @@
-import statistics
 from django.middleware import csrf
 from django.http import Http404
 from django.utils.decorators import method_decorator
@@ -12,7 +11,6 @@ from rest_framework.decorators import action
 from rest_framework import filters
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
-from django.shortcuts import get_object_or_404
 
 from openedx.core.djangoapps.cors_csrf.authentication import SessionAuthenticationCrossDomainCsrf
 from openedx.features.genplus_features.genplus.models import (
@@ -20,8 +18,6 @@ from openedx.features.genplus_features.genplus.models import (
 )
 from openedx.features.genplus_features.genplus.constants import JournalTypes
 from openedx.features.genplus_features.common.display_messages import SuccessMessages, ErrorMessages
-from openedx.features.genplus_features.genplus_learning.models import ClassUnit, ClassLesson
-from openedx.features.genplus_features.genplus_learning.api.v1.serializers import ClassSummarySerializer
 from .serializers import (
     CharacterSerializer,
     ClassSerializer,
@@ -149,22 +145,6 @@ class ClassViewSet(GenzMixin, viewsets.ModelViewSet):
         }
         return Response(data, status=status.HTTP_200_OK)
 
-    def retrieve(self, request, group_id=None):  # pylint: disable=unused-argument
-        """
-        Returns the summary for a Class
-        """
-        gen_class = get_object_or_404(Class, pk=group_id)
-        class_units = ClassUnit.objects.select_related('gen_class', 'unit').prefetch_related('class_lessons')
-        class_units = class_units.filter(gen_class=gen_class)
-        data = ClassSummarySerializer(class_units, many=True).data
-
-        for i in range(len(data)):
-            lessons = data[i]['class_lessons']
-            data[i]['unit_progress'] = round(statistics.fmean([lesson['class_lesson_progress']
-                                                               for lesson in lessons])) if lessons else 0
-
-        return Response(data, status=status.HTTP_200_OK)
-
     @action(detail=True, methods=['put'])
     def add_my_class(self, request, group_id=None):  # pylint: disable=unused-argument
         """
@@ -188,21 +168,6 @@ class ClassViewSet(GenzMixin, viewsets.ModelViewSet):
             teacher_class.save()
             return Response(SuccessMessages.CLASS_REMOVED_FROM_FAVORITES.format(class_name=gen_class.name),
                             status=status.HTTP_204_NO_CONTENT)
-
-    @action(detail=True, methods=['put'])
-    def unlock_lesson(self, request, pk=None):  # pylint: disable=unused-argument
-        """
-       unlock the lesson of the unit
-        """
-        try:
-            lesson = ClassLesson.objects.get(pk=pk)
-        except ClassLesson.DoesNotExist:
-            return Response(ErrorMessages.LESSON_NOT_FOUND.format(pk), status=status.HTTP_404_NOT_FOUND)
-        if not lesson.is_locked:
-            return Response(ErrorMessages.LESSON_ALREADY_UNLOCKED, status.HTTP_204_NO_CONTENT)
-        lesson.is_locked = False
-        lesson.save()
-        return Response(SuccessMessages.LESSON_UNLOCKED, status.HTTP_204_NO_CONTENT)
 
 
 class JournalViewSet(GenzMixin, viewsets.ModelViewSet):
