+++ openedx/features/genplus_features/genplus_learning/api/v1/views.py
@@ -1,3 +1,4 @@
+import statistics
 from django.shortcuts import get_object_or_404
 from rest_framework import generics, status, views, viewsets, mixins, filters
 from rest_framework.permissions import IsAuthenticated
@@ -8,8 +9,8 @@ from openedx.features.genplus_features.genplus.models import GenUser, Student, C
 from openedx.features.genplus_features.common.display_messages import SuccessMessages, ErrorMessages
 from openedx.features.genplus_features.genplus.api.v1.permissions import IsStudentOrTeacher, IsTeacher
 from openedx.features.genplus_features.genplus_learning.models import (Program, ProgramEnrollment,
-                                                                       ClassUnit,)
-from .serializers import ProgramSerializer, ClassStudentSerializer
+                                                                       ClassUnit, ClassLesson,)
+from .serializers import ProgramSerializer, ClassStudentSerializer, ClassSummarySerializer
 
 
 class ProgramViewSet(viewsets.ModelViewSet):
@@ -70,3 +71,43 @@ class ClassStudentViewSet(mixins.ListModelMixin,
         except Class.DoesNotExist:
             return Student.objects.none()
         return gen_class.students.select_related('gen_user__user').all()
+
+
+class ClassSummaryViewSet(mixins.RetrieveModelMixin,
+                          viewsets.GenericViewSet):
+    authentication_classes = [SessionAuthenticationCrossDomainCsrf]
+    permission_classes = [IsAuthenticated, IsTeacher]
+    serializer_class = ClassSummarySerializer
+    queryset = Class.objects.all()
+    lookup_field = 'group_id'
+
+    def retrieve(self, request, group_id=None):  # pylint: disable=unused-argument
+        """
+        Returns the summary for a Class
+        """
+        gen_class = self.get_object()
+        class_units = ClassUnit.objects.select_related('gen_class', 'unit').prefetch_related('class_lessons')
+        class_units = class_units.filter(gen_class=gen_class)
+        data = self.get_serializer(class_units, many=True).data
+
+        for i in range(len(data)):
+            lessons = data[i]['class_lessons']
+            data[i]['unit_progress'] = round(statistics.fmean([lesson['class_lesson_progress']
+                                                               for lesson in lessons])) if lessons else 0
+
+        return Response(data, status=status.HTTP_200_OK)
+
+    @action(detail=True, methods=['put'])
+    def unlock_lesson(self, request, lesson_id=None):  # pylint: disable=unused-argument
+        """
+       unlock the lesson of the unit
+        """
+        try:
+            lesson = ClassLesson.objects.get(pk=lesson_id)
+        except ClassLesson.DoesNotExist:
+            return Response(ErrorMessages.LESSON_NOT_FOUND.format(lesson_id), status=status.HTTP_404_NOT_FOUND)
+        if not lesson.is_locked:
+            return Response(ErrorMessages.LESSON_ALREADY_UNLOCKED, status.HTTP_204_NO_CONTENT)
+        lesson.is_locked = False
+        lesson.save()
+        return Response(SuccessMessages.LESSON_UNLOCKED, status.HTTP_204_NO_CONTENT)
